release(cli): prepare v0.1.6
- add --author flag to `prompt create` (was hardcoded to "unknown")
- skip LLM judge gracefully in offline mode instead of erroring
- fix --since hours filter (12h now correctly filters last 12h, not 1 day)
- fix invalid escape sequence deprecation warning in doctor.py

+++ packages/cli/evalflow/storage/db.py
@@ -215,7 +215,7 @@ class EvalflowDB:
     async def list_runs(
         self,
         limit: int = 20,
-        since_days: int | None = None,
+        since_days: float | None = None,
         failed_only: bool = False,
     ) -> list[dict[str, Any]]:
         """List recent runs with optional filtering."""
